Main screen rework part one Sliders work fine

+++ Common/src/main/java/com/leobeliik/extremesoundmuffler/interfaces/ISoundLists.java
@@ -9,6 +9,6 @@ public interface ISoundLists {
     Set<String> forbiddenSounds = new HashSet<>();
     SortedSet<ResourceLocation> soundsList = new TreeSet<>();
     SortedSet<ResourceLocation> recentSoundsList = new TreeSet<>();
-    Map<ResourceLocation, Float> muffledSounds = new HashMap<>();
+    Map<ResourceLocation, Double> muffledSounds = new HashMap<>();
     List<Anchor> anchorList = new ArrayList<>();
 }
